+++ packages/af-utils-react-virtual-headless/src/hooks/useComponentSubscription/index.js
@@ -8,14 +8,8 @@ const EVT_TO_PROP = ["from", "to", "scrollSize", "sizesHash"];
 const useComponentSubscription = (model, events) => {
     events ||= EVT_ALL;
 
-    /*
-        Naive way is just to combine integers to string.
-        But can do more optimal with low collision rate.
-
-        TODO: test it
-    */
     const getHash = () =>
-        events.reduce((acc, e) => acc ^ (model[EVT_TO_PROP[e]] << e), 31);
+        events.reduce((acc, e) => acc + model[EVT_TO_PROP[e]] + "_", "");
 
     useSyncExternalStore(
         useCallback(listener => model.on(listener, events), [events]),