+++ packages/af-utils-react-mobx-table/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @af-utils/react-mobx-table
 
+## 0.0.31
+
+### Patch Changes
+
+-   improved scrollTo
+-   Updated dependencies
+    -   @af-utils/react-table@0.0.29
+    -   @af-utils/react-virtual-headless@0.0.24
+    -   @af-utils/styled@0.0.11
+
 ## 0.0.30
 
 ### Patch Changes