+++ packages/af-utils-react-table/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @af-utils/react-table
 
+## 0.0.29
+
+### Patch Changes
+
+-   improved scrollTo
+-   Updated dependencies
+    -   @af-utils/react-virtual-headless@0.0.24
+    -   @af-utils/styled@0.0.11
+
 ## 0.0.28
 
 ### Patch Changes