+++ packages/af-utils-styled/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @af-utils/styled
 
+## 0.0.11
+
+### Patch Changes
+
+-   improved scrollTo
+
 ## 0.0.10
 
 ### Patch Changes